servers/memory/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-storage
